디스크 용량 알람
시스템설정에서 값 설정

+++ NotchingGradeInsp/View/CImageDispDlg.cpp
@@ -1136,7 +1136,7 @@ void CImageDispDlg::DrawCross(CDC* pDC)
 
 	CSize size = pBmpDest->GetImgSize();
 
-	int nSize = m_vptDrawCross.size();
+	int nSize = (int)m_vptDrawCross.size();
 	int nCrossSize = 100;
 
 	for( int i=0; i<nSize; i++ )
